140409 by corwin on 2019/08/16 (pair: sclafani) Merged 139742 to linux-vdo [BZ 1659303]. 139742 by corwin on 2019/06/20 (pair: msakai) Clean up slab summary in light of suspend/resume [VDO-4402].
This is a clean up which comes out of the suspend/resume work. Since it
was not strictly necessary to fix the bug, it has been deferred until
after the main change.

This change does 4 things:

1) It eliminates writing out the slab summary zone blocks when
resuming. Now that we are guaranteed that there can be no summary
updates once the summary has been suspended, we know that both the
original and copied summaries are both complete after a grow physical
or failed grow physical, so this write is no longer necessary.

2) It cleans up the handling of slab summary block pbns by always
calculating them when a block is written instead of relying on the
suspend and resume process to make sure that origin changes are
propagated to the individual blocks.

3) It converts checking for drain completeness to use a count instead
of sampling every block.

4) It ensures that all errors attempting to update the slab summary
put the VDO into read-only mode.

.../slabSummaryInternals.h
.../slabSummary.c
Replaced SlabSummaryBlock's "pbn" field with an "index" field to hold
the index of the block in its zone's summary.
Modified initializeSlabSummaryBlock() to take the index of the block in
its zone instead of a PBN; removed the now unused "pbn" parameter
from makeSlabSummaryZone().
Modified launchWrite() to always calculate the PBN at which to write
the block.
Modified the load path to use the summary's origin instead of the
first SummaryZone's first block to determine the PBN from which to
start reading and writing the summary.
Renamed the "currentlyWriting" field to "writing".
Removed the "needsWriting" field. Since the only times we need to write
blocks are either at load time (which didn't use this field), or when
a block has been dirtied by a call to updateSlabSummaryEntry(), the
existence of waiters on the "nextUpdateWaiters" queue is now
equivalent to this field.
Modified launchWrite() to assume it is only called on blocks which
should be written. Reworked the callbacks and error handling from
launchWrite() to reflect the changes in fields and operations.
Added a "writeCount" field to SlabSummaryZone to count the number of
blocks out for writing.
Modified checkForOperationComplete() to wait for the "writeCount" to be
0 rather than checking the state of each block explicitly.
Removed the triggering of writes in drainSlabSummaryZone() and
resumeSlabSummaryZone().
Removed launchWriteOfAllBlocks().

.../refCounts.c
Modified finishSummaryUpdate() to enter read-only mode on error.

.../slabJournal.c
Modified releaseJournalLocks() to not log errors if the error code is
VDO_READ_ONLY since such logging would be both superfluous and prone
to flooding.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>