140345 by tomj on 2019/08/12 (pair: sweettea) UDSUP-7 Merge change 140135 to Jasper 140135 by tomj on 2019/07/22 (pair: sweettea) UDSUP-7 Make and test a dm-bufio client.
Note that with this change, we will no longer be able to build
linux-uds on RHEL7.

.../ioFactory.h
Add the makeBufio method to the kernel version of an IOFactory.  This
method makes a dm_bufio_client so that UDS code can access storage
using the dm-bufio facility.
Bring ioFactory.h closer into line with our coding conventions,
especially with the location of the warn_unused_result attribute.

.../request.h
.../opaqueTypes.h
Change the name of struct request.  This is to resolve a naming
conflict with the linux kernel structure of the same name.

140417 by tomj on 2019/08/19 (pair: sclafani)
UDSUP-7 Merge change 140147 to Jasper
140147 by tomj on 2019/07/23 (pair: sweettea)
UDSUP-7 Use dm-bufio to read and write the volume

Pair:sweettea

.../volume.h
.../volume.c
.../volumeInternals.c
In kernel mode, the Volume needs to have a dm_bufio_client.  In user
mode, the Volume needs an IORegion and a scratch page.
Using #ifdef __KERNEL__, make the kernel mode code read and write the
volume using dm-bufio, and make the user mode code perform I/O to the
volume using an IORegion.
Kernel mode writes will use the memory buffers provided by dm-bufio.
Kernel mode reads will copy the data from the dm-bufio buffers into
our UDS buffers.
In the methods that use dm-bufio, use the name "buffer" for the
dm-bufio buffer control block, and use the name "data" for the
page of data.
Make the index page data an argument to the donateIndexPageLocked
method.
There are some log messages that use logError in the index page code
path and use logWarning in the record page code path.  Make them both
use logWarning.
There is a message about "no delta lists packed", which probably is of
little interest to users of the index.  Change it from using logInfo
to using logDebug.

.../indexLayout.h
.../indexLayout.c
Create the openVolumeBufio method for opening the volume using dm-bufio
in kernel mode.  Make the openVolumeRegion method visible only in
user mode.

140434 by tomj on 2019/08/19 (pair: jwiele)
UDSUP-7 Merge change 140174 to Jasper
140174 by tomj on 2019/07/24 (pair: msakai,sweettea)
UDSUP-7 Remove the error return from putIOFactory.

Freeing an IOFactory returns an error code, but there is not much use
that can be made of the error code.  The easiest approach looks like
not having the error code returned at all.

In kernel mode, there actually are no errors that can be returned.
The system services that we use all have a void return.  So we always
return UDS_SUCCESS.  That works, but when we upstream this code we
will be asked why we have a return status that is always success.

In user mode, the one system service that can return an error is
the close system call.  The "man 2 close" page has some interesting
comments on the errors that can be returned.  The interesting errors
that can happen in production are all delayed reports of device errors.
The trouble is that there is nothing we can do about those errors other
than log that they occurred.  The file descriptor is always closed, and
all we know is that some error happened.

The man page suggests that a better approach is use the fsync system
call before calling the close system call.  It gives us a better chance
to actually do something.  And in fact, whenever we finish writing a
section of the index, we always do call fsync.  So the close error that
we cannot deal with should never occur.

Our usage of the close system call is wrapped by the loggingClose
method, so any close error will send a message to the log.

.../ioFactory.h
.../ioFactoryLinuxUser.c
The putIOFactory method does not return a value.

.../indexLayout.c
.../indexLayoutLinuxUser.c
.../fileIORegion.c
Remove checking the return value from putIOFactory.

.../fileUtils.h
Remove the warn_unused_result attribute from the closeFile method.

140440 by corwin on 2019/08/19 (pair: sclafani)
Merged 140180 to linux-vdo.
140180 by corwin on 2019/07/24 (pair: sweettea)
Simplify exiting of recovery mode [VDO-4430].

This change is a precursor to the main work of changing the vdo load
process to not write while in the constructor. That work entails moving
the bulk of the work currently done in the load operation into the
resume operation which DM will invoke once the constructor returns.

One impediment to moving that work is that there will no longer be an
explicit call to scrub unrecovered slabs once the VDO is ready to come
online. As such, the current entry point where we register the callback
which is used to call leaveReadOnlyMode() will no longer exist.

This change makes the VDO's state atomic and passes a pointer to that
state into the SlabDepot so that the depot can handle leaving recovery
mode on its own. This has the additional benefit of allowing for the
removal of the scrubbing completion. The use of the scrubbing completion
was only correct by happenstance since it is allocated as part of making
the slab depot, and then when scrubbing completes on all zones, it is
sent to the admin thread in order to call leaveRecoveryMode() (which
prior to this change had to be called from the admin thread). However,
once that completion is enqueued for the admin thread, nothing in the
depot is ever notified of when the callback has been processed and the
completion is no longer in use. As it happens, there is no
use-after-free race here, but only because the slab depot must be
drained before it can be freed, and the suspend operation which
initiates the drain must process the suspend completion back on the
admin thread in order to know that the depot has drained. If the
scrubbing completion were enqueued for the admin, that must happen
before the last zone which was scrubbing has been drained, and
therefore, the scrubbing completion must be processed before the
continuation of the suspend operation. This relies on the fact that
there is no way for completions enqueued from the same producer thread
onto any given queue can get reordered. By removing the use of the
scrubbing completion, the need for this subtle correctness guarantee
goes away.

.../vdoInternal.h
.../vdo.c
Changed the VDO's 'state' field to be an Atomic32; added getVDOState()
and setVDOState() methods to access it.
Removed leaveRecoveryMode().

.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
Added a 'vdoState' field to SlabDepot and an associated parameter to
both makeSlabDepot() and decodeSlabDepot().
Removed the 'scrubbingCompletion' field from SlabDepot.
Removed all of the parameters used to prepare the scrubbing completion
from scrubAllUnrecoveredSlabs().
Modified notifyZoneFinishedScrubbing() to update the VDO's state and
log when leaving recovery mode itself rather than launching the
scrubbing completion.

.../vdo.c
.../upgrade.c
.../vdoLoad.c
.../vdoRecovery.c
.../vdoResume.c
.../vdoSuspend.c
.../vdoConfig.c
Updated all methods which access the VDO's state to do so through
getVDOState() and setVDOState().

.../vdoLoad.c
Reflected change to scrubAllUnrecoveredSlabs().
Reflected change to decodeSlabDepot().

.../vdoConfig.c
Reflected change to makeSlabDepot().

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>