140693 by tomj on 2019/09/11 (pair: sweettea) UDSUP-5 Merge change 140449 to Jasper 140449 by jwiele on 2019/08/20 (pair: raeburn) UDSUP-5 Remove unused grid only file functions.
.../fileUtils.h
.../fileUtils.c
Remove wrapper functions for the fcntl(2) and mkdir(2) system
calls that are only relevant to grid mode which is not
supported in linux-uds or linux-vdo.

140694 by tomj on 2019/09/11 (pair: sweettea)
UDSUP-7 Merge change 140654 to Jasper
140654 by tomj on 2019/09/05 (pair: sweettea)
UDSUP-7 Renamed the struct cachedPage fields.

.../pageCache.h
.../pageCache.c
.../volume.c
Add the "cp_" prefix to all the fields of struct cachedPage.  This
makes it easier to use emacs, grep and gid to modify the pageCache
code.

140712 by jshimkus on 2019/09/11 (pair: sweettea)
Merge 140502 to linux-vdo.
140502 by jshimkus on 2019/08/22 (pair: bjohnston)
[VDO-4730] Warn on vdo start/stop when VDO already started/stopped.

[BZ 1738651]
vdo manager was checking if the VDO instance was already running
without regard to the possibility that user requested a forced
rebuild.  This resulted in a successful completion result of the start
without actually performing the forced rebuild providing a misleading,
at best, indication to the user.  This changeset modifies the existing
VDO already started message to include a clause indicating no changes
were made and now logs the message at warning-level.

The changeset also changes the pre-existing VDO already stopped message
from info-level to warning-level to provide consistency for all
start/stop scenarios.  In all cases the actual start/stop exits with a
success status.

.../VDOService.py
Changed existing VDO already started/stopped messages to warning-level.
Added 'no changes made' clause to the VDO already started message.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: bjohnston <bjohnsto@redhat.com>