140763 by tomj on 2019/09/16 (pair: sweettea) UDSUP-7 Merge change 140682 to Jasper 140682 by tomj on 2019/09/10 (pair: sweettea) UDSUP-7 Move volumeInternals code into the volume module.
.../volume.h
.../volume.c
.../volumeInternals.h
.../volumeInternals.c
Move the mapToPhysicalPage and allocateVolume methods from the
volumeInternals module to the volume module.  These methods benefit
from compiler inlining when compiled in the volume module, and the
methods that remain in the volumeInternals module will no longer be
needed when we finish using dm-bufio to manage the volume's buffer
space.

140780 by tomj on 2019/09/17 (pair: sweettea)
UDSUP-7 Merge change 140691 to Jasper
140691 by tomj on 2019/09/11 (pair: sweettea)
UDSUP-7 Create struct volume_page

.../volumeStore.h
.../volumeStore.c
Create struct volume_page to be the actual memory buffers used for
index and record pages, and the various methods for reading and
writing the data.  Not all of these methods are used in this
changeset, but they will all be used in the upcoming UDSUP-7
changes.

140781 by tomj on 2019/09/17 (pair: sweettea)
UDSUP-6 Merge change 140692 to Jasper
140692 by tomj on 2019/09/11 (pair: sweettea)
UDSUP-6 RequestQueue refactoring preparing to use kernel wait queues

Mostly this change is moving some code around and a little refactoring
so that the user and kernel mode implementations will look similar.

.../requestQueue.c
Refactor the makeRequestQueue and initializeRequestQueue methods into
a single method.
Add a memory barrier to the end of the makeRequestQueue method, so that
all initialization of the RequestQueue will be visible to any user of
the RequestQueue.
Create the wakeUpWorker helper method.  The contents of this method
will be different in the kernel mode implementation, but the callers
(requestQueueEnqueue and requestQueueFinish) will be nearly
identical.
Change the error logging in requestQueueFinish to warning logging.  In
practise, this is an error that we cannot do anything about, and it
can happen only in user mode.  The joinThreads method in kernel mode
always returns UDS_SUCCESS.

140782 by tomj on 2019/09/17 (pair: sweettea)
UDSUP-7 Merge change 140722 to Jasper
140722 by tomj on 2019/09/12 (pair: sweettea)
UDSUP-7 Use a struct volume_page for the scratch page.

.../volume.h
.../volume.c
Use a struct volume_page for the scratch page.
Refactor the writing of an index or record page to use the
prepareToWriteVolumePage and writeVolumePage methods.
When we finish writing a chapter, call the releaseVolumePage method.
In kernel mode, this will give up the reservation the last page that
was written.

140806 by tomj on 2019/09/18 (pair: sclafani)
UDSUP-7 Merge change 140756 to Jasper
140756 by tomj on 2019/09/16 (pair: sweettea)
UDSUP-7 Use a struct volume_page for the page cache.

.../pageCache.h
.../pageCache.c
Use a struct volume page for a page in the page cache.

.../volume.h
.../volume.c
Refactor the reading and usage of CachePages to use the volume_page
interfaces.
When reading the index pages for an entire chapter, prefetch the entire
set of pages.
When donating a newly written index page to the page cache, replace
copying the data with just swapping the pointers to the data.
Delete the offsetForChapter method, which is no longer used.
When probing a chapter in preparation for replaying the volume,
prefetch the entire chapter index.
Reserve enough buffers in dm-bufio to hold the contents of the page
cache.

.../volumeInternals.c
.../volumeInternals.h
Delete the volumeInternals module.

140807 by tomj on 2019/09/18 (pair: sclafani)
UDSUP-7 Merge change 140758 to Jasper
140758 by tomj on 2019/09/16 (pair: sweettea)
UDSUP-7 Better detection of dm-bufio write errors

Byte streams written using a BufferedWiter have been depending upon
the dm_bufio_client_destroy method flushing any outstanding writes to
the storage device.  This causes us to lose any errors reported by
those writes.  Add an explicit call to the dm_bufio_write_dirty_buffers
method and process any resulting errors.

.../bufferedWriter.c
Call the dm_bufio_write_dirty_buffers and deal with any resulting
errors.

140808 by tomj on 2019/09/18 (pair: sclafani)
UDSUP-7 Merge change 140761 to Jasper
140761 by tomj on 2019/09/16 (pair: sweettea)
UDSUP-7 Fix the Configuration_n1 test on vfarm-114

The UDSUP-7 work has introduced a problem that we see when we run the
Configuration_n1 test on vfarm-114.  The reason that vfarm-114 is
special is that it is the only FEDORA29 vfarm that has enough memory to
suggest that a 2GB index will be able to run, but does not have a large
enough storage device for a 2GB sparse index.  With the dm-bufio work
in place, we do not cleanly detect the storage errors.

.../indexLayout.h
.../indexLayout.c
Rename the size argument to the makeIndexLayoutFromFactory method.  The
new name is namedSize, which reflects that fact that this is the size
of the index specified in the index name string.  Note that VDO
always specifies the namedSize in its index name string.
Compare the namedSize to the actual size of the block device.  If the
block device is smaller, return a UDS_INSUFFICIENT_INDEX_SPACE error.
If the namedSize is smaller is smaller, use the smaller size as the
size of the device.
If we are making an IndexLayout for a udsCreateLocalIndex call, a
UdsConfiguration will be supplied.  Compute the size of the index
that results from this configuration.  If the block device is smaller
than the index requirement, return a UDS_INSUFFICIENT_INDEX_SPACE
error.  Otherwise use the size needed for the index configuration.
Remove the size alignment check in the creatIndexLayout method.
This check will alway pass because the size returned from the
udsComputeIndexSize method is always correctly aligned.
Remove the recomputeSizes call from the createIndexLayouMethod.  VDO
will never need this code path, and in tests it causes us to use the
entire extra device space for additional save slots.

.../ioFactory.h
.../ioFactoryLinuxUser.c
Create the getWritableSize method.  In the kernel mode implementation,
this method will return the size of the block device.  In the user
mode implementation, we will return the largest conceivable size that
Linux supports.

140812 by tomj on 2019/09/18 (pair: sclafani)
Fix side effects of Jasper merge 140806

.../Makefile
Delete the volumeInternals module from the kernel build.  This module
was deleted by merge 140806.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>