140831 by tomj on 2019/09/20 (pair: sweettea) ALB-2990 Merge change 140762 to Jasper 140762 by tomj on 2019/09/16 (pair: sweettea) ALB-2990 Merge change 140723 to linux-uds 140723 by tomj on 2019/09/12 (pair: sweettea) ALB-2990 Do not rebuild the index if loading had an allocation error
.../index.c
If loading an index fails with an ENOMEM error, then we had a memory
allocation error.  In this case, do not try to rebuild the index.

140832 by tomj on 2019/09/20 (pair: sweettea)
UDSUP-7 Merge change 140779 to Jasper
140779 by tomj on 2019/09/17 (pair: sweettea)
UDSUP-7 Use struct volume_pages for the sparse chapter cache.

.../cachedChapterIndex.h
.../cachedChapterIndex.c
Replace the single buffer containing the entire chapter index with an
array of volume_page structs, where each struct volume_page holds a
single index page.

.../volume.h
.../volume.c
Change the readChapterIndexFromVolume method to populate the array of
struct volume_pages.
Include the buffers needed for the sparse chapter cache in the number
of reserved buffers for the volume store.

140839 by tomj on 2019/09/20 (pair: sweettea)
UDSUP-7 Merge change 140778 to Jasper
140778 by tomj on 2019/09/17 (pair: sweettea)
UDSUP-7 Use prefetch when replaying the volume

.../index.c
Add code to the replayVolume method that prefetches the entire chapter
that is about to be replayed.

140863 by tomj on 2019/09/24 (pair: sweettea)
ALB-3021 Merge change 140810 to Jasper
140810 by tomj on 2019/09/18 (pair: sclafani)
ALB-3021 Merge change 140574 to linux-uds
140574 by tomj on 2019/08/28 (pair: sweettea)
ALB-3021 Add ChapterIndexPage fields

The real problem in ALB-3021 is that we have been writing the fields of
the DeltaPageHeader in native endian byte order, and therefore
transferring an index between hosts of differing endianness does not
work.

This changeset makes copies of the header field values from the
DeltaPageHeader.  The copies will live in the ChapterIndexPage.  These
copies replace getter methods that read the values directly from the
DeltaPageHeader in the index page.  This is a safe transformation
because the volume pages are read-only once they have been written.

A future change will be reading the values from the DeltaPageHeader in
an endianness appropriate fashion.

.../chapterIndex.h
.../chapterIndex.c
Add the lowestListNumber, highestListNumber and virtualChapterNumber
fields to the ChapterIndexPage.  Set these new fields in the
initializeChapterIndexPage method.
Delete the getChapterIndexVirtualChapterNumber,
getChapterIndexLowestListNumber and getChapterIndexHighestListNumber
methods.

.../chapterIndex.c
.../index.c
.../volume.c
Replace getChapterIndexVirtualChapterNumber method calls with direct
references to the virtualChapterNumber field of a ChapterIndexPage.
Replace getChapterIndexHighestListNumber and
getDeltaIndexHighestListNumber method calls with direct
references to the highestListNumber field of a ChapterIndexPage.
Replace getChapterIndexLowestListNumber and
getDeltaIndexLowestListNumber method calls with direct
references to the lowestListNumber field of a ChapterIndexPage.

140867 by tomj on 2019/09/24 (pair: sweettea)
ALB-3021 Merge change 140827 to Jasper
140827 by tomj on 2019/09/20 (pair: sweettea)
ALB-3021 Merge change 140683 to linux-uds
140683 by tomj on 2019/09/10 (pair: sweettea)
ALB-3021 Rename ChapterIndexPage to DeltaIndexPage.

.../chapterIndex.h
.../chapterIndex.c
.../deltaIndex.h
.../deltaIndex.c
Rename the ChapterIndexPage to the new name DeltaIndexPage.  They
really are the same thing, and it is the deltaIndex code that really
knows about the internals of a DeltaIndexPage.  Move the structure
definition to the deltaIndex module.
Set the header fields of the DeltaIndexPage in the
initializeDeltaIndexPage method.
Eliminate the getter methods for the DeltaPageHeader fields.  The
initializeDeltaIndexPage method loads these values directly into
the DeltaIndexPage structure.

.../cachedChapterIndex.h
.../cachedChapterIndex.c
.../index.c
.../pageCache.h
.../pageCache.c
.../sparseCache.c
.../volume.h
.../volume.c
Do the ChapterIndexPage to DeltaIndexPage rename.

140871 by tomj on 2019/09/24 (pair: sweettea)
Use the system error code ERANGE in the vdo user mode code

.../parseUtils.c
Replace the usage of the UDS error code UDS_OUT_OF_RANGE with the
system error code ERANGE.  This will be more useful as the callers
of parseIndexConfig reports the error using the errx method.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>