140910 by tomj on 2019/09/25 (pair: sweettea) UDSUP-6 Merge change 140829 to Jasper 140829 by tomj on 2019/09/20 (pair: sweettea) UDSUP-6 Reimplement RequestQueue to use struct wait_queue.
The Linux kernel provides a wait_queue facility to implement the basic
operation of a thread waiting for a condition.  Albireo implemented
this operation using an EventCount, which avoids locking as much as
possible, only using a semaphore when it is necessary to wait.  When
it uses this semaphore, its performance depends upon the ability to
wait for a semaphore with a timeout that has microsecond granularity.

It is not difficult to implement such a semaphore in Linux, but past
attempts to add such a semaphore to the kernel have met with failure.

But the good news is that the wait_queue facility has the same
properties that an EventCount has.  It avoids locking as much as
possible, has the same prepare/wait/cancel semantics, and it has a
timeout that uses a high resolution timer.

.../requestQueueUser.c
.../requestQueue.c
User mode UDS will continue to use the existing
RequestQueue implementation.  Branch src/uds/requestQueue.c to
userLinux/uds/requestQueueUser.c, and delete src/uds/requestQueue.c.

.../Makefile
Remove requestQueue.c from the list of UDS common source files.
Add requestQueueKernel.c to the kernel build, and requestQueueUser.c
to the user mode build.

140911 by tomj on 2019/09/25 (pair: sweettea)
ALB-3021 Merge change 140834 to Jasper
140834 by tomj on 2019/09/20 (pair: sweettea)
ALB-3021 Merge change 140707 to linux-uds
140707 by tomj on 2019/09/11 (pair: sweettea)
ALB-3021 Improve chapter index verification during rebuild

.../index.c
When rebuilding the IndexPageMap, verify that the chapter index pages
contain a consistent number of index lists.  Due to ALB-3021, it is
conceivable but very rare for an inconsistent number to happen.  The
resulting errors would be confusing.

140914 by tomj on 2019/09/25 (pair: sweettea)
ALB-3021 Merge change 140841 to Jasper
140841 by tomj on 2019/09/20 (pair: sweettea)
ALB-3021 Merge change 140724 to linux-uds
140724 by tomj on 2019/09/12 (pair: sweettea)
ALB-3021 Refactor initializing a DeltaIndexPage

.../deltaIndex.c
Refactor the initializeDeltaIndexPage method, splitting out the
verifyDeltaIndexPage method.  This will be a preferred refactoring
when we have to read a chapter index page that may be either little
or big endian.  In that case, the initializeDeltaIndexPage method may
need to call the verifyDeltaIndexPage method twice.
Do less logging in verifyDeltaIndexPage.  The log messages, while
detailed and informative during development, are not of great use
to end users.  If we try to verify an index page using the wrong
endianness, they will put useless noise into the kernel log.  And
it is likely that the upstream community will disapprove of the
logging.
Do not skip the check of the expected nonce when the expected nonce is
zero.  We only used a zero nonce during testing or for the old-style
multi-file Albireo user mod index.
Remove the redundant check that the last list ends on the page, because
it is covered by the check that the last list ends while leaving room
on the page for the post-field guard bits.

140917 by joe on 2019/09/25 (pair: sweettea)
Detypedef blockMapTree.h/blockMapTreeInternals.h (VDOUP-6).

.../blockMapTree.h
.../blockMapInternals.h
Remove struct typedef(s).
Decamelcase the struct identifier(s).

.../blockMapTree.c
.../blockMapTreeInternals.h
.../forest.c
.../forest.h
Adapt to the above changes.

140923 by tomj on 2019/09/25 (pair: sweettea)
Fix side effects of Jasper merges 140910 and 140912.

.../memoryUsage.c
Modify the arguements to the getMemoryStats method to match the
interface changes in merge 140912.

.../Makefile
Reflect the renaming and moving of files in merge 140910.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>